+++ src/soot/util/Numberer.java
@@ -27,9 +27,9 @@ public interface Numberer {
     /** Should return the number that was assigned to object o that was
      * previously passed as an argument to add().
      */
-    public int get( Object o );
+    public long get( Object o );
     /** Should return the object that was assigned the number number. */
-    public Object get( int number );
+    public Object get( long number );
     /** Should return the number of objects that have been assigned numbers. */
     public int size();
 }
